refactor: separate business logic from NavBar component
