+++ projects/sasil-web/src/components/molelcules/SearchBar/index.ts
@@ -1,5 +1,7 @@
 import SearchBar from './SearchBar';
+import SearchBarWrapped from './SearchBarWrapped';
 
 export default SearchBar;
+export { SearchBarWrapped };
 
 export type { SearchBarProps } from './SearchBar';
