+++ projects/sasil-web/src/components/templates/NavBar/NavBar.stories.tsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import NavBar from '.';
+import NavBar, { NavBarProps } from '.';
 
 export default {
   title: 'components/template/NavBar',
   component: NavBar,
 } as ComponentMeta<typeof NavBar>;
 
-const Template: ComponentStory<typeof NavBar> = () => <NavBar />;
+const Template: ComponentStory<typeof NavBar> = ({ children }: NavBarProps) => (
+  <NavBar>{children}</NavBar>
+);
 
 export const Default = Template.bind({});
