+++ projects/sasil-web/src/components/templates/NavBar/index.ts
@@ -1,3 +1,4 @@
 import NavBar from './NavBar';
 
 export default NavBar;
+export type { NavBarProps } from './NavBar';
